Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dotenv==0.9.9
easyocr==1.7.2
fastapi==0.127.1
filelock==3.20.1
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
PySocks==1.7.1
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.26.0
scipy==1.16.3
selenium==4.39.0
setuptools==80.9.0
shapely==2.1.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
sympy==1.14.0
tifffile==2025.12.20
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
